HXPController::poll(): Set status on all axes
The following seemes to have introduced a small regression:
commit 808e26ef1a8418d8d07742454a6c311e77d5654a
Moved the polling from HXPAxis::poll to HXPController::poll

The motorStatusProblem and motorStatusPowerOn are only set
on the first axis.
Restore the old behaviour and set the bits on all axes in
HXPController::poll()

To be done in the next commit:
Set the encoder and motor position here as well.