Add is_firmware_hxpd_
Add a flag is_firmware_hxpd_ in HXPDriver.cpp which switches the code
in poll() to take a different look at the state.
In our device I think that the power is always on.
Treat the homed stated differently: state 43 is homing.
Which means that there is (still) a problem.

Forward is_firmware_hxpd_ into HXPSetHexapodForFirmwareXPS_D()
so that we can, in theory, keep the driver running and connect
to a different controller.
Even if this is not happening so often, the code is now cleaner.

Changes to be committed:
    modified:   src/HXPDriver.cpp
    modified:   src/HXPDriver.h
    modified:   src/hxp_drivers.cpp
    modified:   src/hxp_drivers.h